First version of the code to perform fits
The code fits some data points to a frequentist distribution, plots the error band from the fit, obtained using the derivatives of the fitting function with respect to the fitting parameters and using the covariant matrix, and plot the ratio between the data points and fit. The error on the ratio points, result from adding the errors in quadrature from the band and the statistical errors from the data points. The code prints the x-value with the corresponding efficiency and also the value of the ratio and the total error.